+++ docs/.vitepress/theme/index.ts
@@ -5,6 +5,7 @@ import DemoGkAlert from '../components/demos/alert/DemoGkAlert.vue'
 import DemoGkButton from '../components/demos/button/DemoGkButton.vue'
 import DemoGkMenu from '../components/demos/containment/menu/DemoGkMenu.vue'
 import DemoGkTabs from '../components/demos/navigation/tabs/DemoGkTabs.vue'
+import DemoGkPagination from '../components/demos/navigation/pagination/DemoGkPagination.vue'
 import DemoGkTooltip from '../components/demos/containment/tooltip/DemoGkTooltip.vue'
 import DemoGkBottomSheet from '../components/demos/containment/bottom-sheet/DemoGkBottomSheet.vue'
 import DemoGkDialog from '../components/demos/containment/dialog/DemoGkDialog.vue'
@@ -37,6 +38,7 @@ export default {
     app.component('DemoGkSelect', DemoGkSelect)
     app.component('DemoGkMenu', DemoGkMenu)
     app.component('DemoGkTabs', DemoGkTabs)
+    app.component('DemoGkPagination', DemoGkPagination)
     app.component('DemoGkTooltip', DemoGkTooltip)
     app.component('DemoGkBottomSheet', DemoGkBottomSheet)
     app.component('DemoGkDialog', DemoGkDialog)