+++ src/vue/index.ts
@@ -69,3 +69,15 @@ export { default as GkTabsWindow } from './components/navigation/tabs/GkTabsWind
 export { default as GkTabsWindowItem } from './components/navigation/tabs/GkTabsWindowItem.vue'
 export { parseGkTabItems } from './components/navigation/tabs/gk-tab-items'
 export type { GkTabItem, GkTabItemParsed } from './components/navigation/tabs/gk-tab-items'
+
+export { default as GkPagination } from './components/navigation/pagination/GkPagination.vue'
+export {
+  buildPaginationRange,
+  createRange,
+  getMax,
+} from './components/navigation/pagination/pagination-range'
+export type { PaginationRangeItem } from './components/navigation/pagination/pagination-range'
+export type {
+  GkPaginationControlSlotProps,
+  GkPaginationItemSlotProps,
+} from './components/navigation/pagination/GkPagination.vue'